Report all tag permutations
When creating a custom dashboard in AppSignal, every permutation of a
filter value will be drawn as a separate line - there's no way of
aggregating them. It is, however, possible to filter out specific tag
values.

This change adds reporting of all tag permutations of a metric, with a
custom tag value "any". Using this value in a dashboard, we can draw
metrics aggregating all values of a given tag.